fix lại font chữ, bộ lọc, cơ bản đã done cả 2 giao diện
